+++ archetype/simple-service/src/main/resources/archetype-resources/api/src/main/java/api/HelloWorldResource.java
@@ -5,9 +5,9 @@ package ${package}.api;
 
 import ${package}.api.model.HelloWorldDto;
 import com.cerner.beadledom.jaxrs.GenericResponse;
-import com.wordnik.swagger.annotations.Api;
-import com.wordnik.swagger.annotations.ApiOperation;
-import com.wordnik.swagger.annotations.ApiResponse;
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiResponse;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;