+++ archetype/simple-service/src/main/resources/archetype-resources/api/src/main/java/api/model/HelloWorldDto.java
@@ -6,8 +6,8 @@ package ${package}.api.model;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.base.Preconditions;
-import com.wordnik.swagger.annotations.ApiModel;
-import com.wordnik.swagger.annotations.ApiModelProperty;
+import io.swagger.annotations.ApiModel;
+import io.swagger.annotations.ApiModelProperty;
 
 @ApiModel(value = "HelloWorldDto")
 public class HelloWorldDto {