+++ swagger2/src/test/java/com/cerner/beadledom/swagger2/ChildResource.java
@@ -1,0 +1,4 @@
+package com.cerner.beadledom.swagger2;
+
+public class ChildResource extends ParentResource {
+}