+++ swagger2/src/test/java/com/cerner/beadledom/swagger2/InterfaceResource.java
@@ -1,0 +1,7 @@
+package com.cerner.beadledom.swagger2;
+
+import javax.ws.rs.Path;
+
+@Path("/interface")
+public interface InterfaceResource {
+}